Make sure we have module with no orphans
