--- src/Instances.hs
@@ -1,19 +1,0 @@
-module Instances where
-
-import Algebras
-
-instance Algebras.SemiGroup [a] where
-  (|+|) = (++)
-  {-# INLINE (|+|) #-}
-
-instance Algebras.Monoid [a] where
-  zero = []
-  {-# INLINE zero #-}
-
-instance Algebras.SemiGroup Int where
-  (|+|) = (+)
-  {-# INLINE (|+|) #-}
-
-instance Algebras.Functor [] where
-  fmap f = foldr (\a xs -> f(a) : xs) []
-
